Fixed signed/unsigned bug in Range#to_gv (sum of int and size_t is unsigned).
git-svn-id: svn+ssh://rubyforge.org/var/svn/rb-gsl/trunk/rb-gsl@36 6e764f74-f39f-46f8-8c54-8307d62afe8d

+++ ext/vector_double.c
@@ -330,7 +330,7 @@ VALUE rb_gsl_range_to_gv(VALUE obj)
   if (RTEST(rb_ivar_get(obj, rb_gsl_id_excl))) n = en - beg;
   else n = en - beg + 1;
   v = gsl_vector_alloc(n);
-  for (i = 0; i < n; i++) gsl_vector_set(v, i, beg + i);
+  for (i = 0; i < n; i++) gsl_vector_set(v, i, beg + (int)i);
   return Data_Wrap_Struct(cgsl_vector, 0, gsl_vector_free, v);
 }
 gsl_vector_view* gsl_vector_view_alloc()
